Proprly format dates iso8601 in json output with 'Z' suffix
Was done before but got lost where along the line. From email
thread back in December regarding ExtJS not being able to read
the output of toJson dates without the Z at the end of the old
datetime output:

On Dec 31, 2008, at 3:59, Bob Parker wrote:

combust/lib/Combust/RoseDB/Object/toJson.pm is doing the "to dates" conversions.

IIRC then we have a db toggle to specify the timezone; the toJson module
could look at the same switch and add the Z when it's set.

No, this has nothing to do with the timezone in the DB. At the time we
translate to json it is just a DateTime object. The format you see is the
result of the code below.

if ($v->isa('DateTime')) {
    my $meth = $date_formatter{$column->type};
    $v = $v->$meth; # strftime($date_format{$column->type});
}

I think we should change that to

if ( $v->isa('DateTime') ) {
    my $meth = $date_formatter{ $column->type };
    $v = $v->clone->set_time_zone('UTC')->$meth;
    $v .= "Z" if $meth eq 'iso8601';
}

Thoughts?

Graham.

Works for me.

- ask