Combust::JSON - a subclass of JSON::XS which decodes/encodes certain classes
This is to replace the current hacks in search-server
and ad-server to deal with HTML::TagCloud objects
when going through JSON intermediate form.

First named YP::JSON, the package was renamed because
the code was not tied to YellowBot codebase and
maybe is generic enough to be used in other places.