replace spaces by '+' in memcached keys.
memcached keys can't have space in them.
So this change puts a fix in place so that
every space is replaced by '+'. (I favoured
'+' instead of '-' or '_', because these later
can be more common in usage and this replacement
has a precedent in URL escaping.)
So "san francisco" will match "san+francisco".

The issue with spaces in memcached keys is
due to the text nature of memcached protocol which
uses spaces for delimiting fields in requests.
This can be seen at the implementation of "sub _set"
in Cache::Memcached
(http://cpansearch.perl.org/src/BRADFITZ/Cache-Memcached-1.26/lib/Cache/Memcached.pm)
where the first request line of a STORE command is

    $cmdname $key $flags $exptime $len

A $key with spaces will cause confusion with the
following fields ($flags, $exptime, $len) which
in practice means that the (key, value) pair will not
be stored in cache as expected.