Combust::JSON behavior can be customized with JSON_freeze/JSON_thaw methods
JSON cannot encode CODE refs or other Perl things
like GLOBs. However it is possible to define
operations that exclude such fields before encoding
and restore them after decoding.

This is what this change is for: with the definition
of JSON_freeze (for encoding) and JSON_thaw (for decoding)
methods, Combust::JSON can support that.

The rule to be obeyed by these methods is that an object
of the class being handled

    $obj

should be equivalent to

    $obj->can('JSON_thaw')->( $obj->JSON_freeze )