Add accessor to access raw json value
setup_json_columns adds triggers so the accessor automatically
inflates/deflates the value. now we have a method with _json
suffix which allows access to the encoded value directly

CAVEAT: when RDBO inflates a column it discards the value read
from the DB. this will happen when the accessor is first called
after reading from DB. In this case the _json method will deflate
the column. This JSON value may not be an identical string to
what is stored as the JSON is no in a canonical form