Change Combust::JSON implementation - amenable to inheritance
With the previous approach, we had problems in
decoding/encoding classes whose super classes
were also handled by Combust::JSON.
This changes the implementation to support that.

The most important change is that JSON_thaw() method
should pick the object class in "__class__" key
and probably using the stored value to instantiate
the correct kind of object.